refactor: address PR review comments
- Move serialization.ts to infrastructure/ (not storage-specific)
- Inline fileIO into options.tsx composition root (single consumer)
- Split OptionsApp into SettingsSection and DataSection components
- Use Result type for onExport/onImport instead of plain string
- Clean up options.html static HTML duplicated by Preact render
