--- src/interface/ui/fileIO.ts
@@ -1,23 +1,0 @@
-export const downloadAsJson = (data: unknown, filename: string): void => {
-  const blob = new Blob([JSON.stringify(data, null, 2)], { type: "application/json" });
-  const url = URL.createObjectURL(blob);
-  const a = document.createElement("a");
-  a.href = url;
-  a.download = filename;
-  a.click();
-  URL.revokeObjectURL(url);
-};
-
-export const readFileAsJson = <T>(file: File): Promise<T> =>
-  new Promise((resolve, reject) => {
-    const reader = new FileReader();
-    reader.onload = () => {
-      try {
-        resolve(JSON.parse(reader.result as string));
-      } catch {
-        reject(new Error("Invalid JSON"));
-      }
-    };
-    reader.onerror = () => reject(new Error("Failed to read file"));
-    reader.readAsText(file);
-  });
