+++ src/interface/ui/options.html
@@ -10,16 +10,7 @@
   <header>
     <h1>PluckVocab Options</h1>
   </header>
-  <main>
-    <section>
-      <label for="storage-backend">Storage Backend</label>
-      <select id="storage-backend">
-        <option value="local">Local Storage</option>
-      </select>
-    </section>
-    <button id="save-btn">Save</button>
-    <p id="status"></p>
-  </main>
+  <main></main>
   <script type="module" src="options.js"></script>
 </body>
 </html>
