+++ src/interface/ui/options.tsx
@@ -1,14 +1,37 @@
 import { render } from "preact";
-import { isOk } from "../../domain/index.js";
-import { createStorage } from "../../infrastructure/storage/index.js";
+import { err, isOk, ok, type Result } from "../../domain/index.js";
 import {
   deserializeVocab,
   type SerializedEntry,
   serializeVocab,
-} from "../../infrastructure/storage/serialization.js";
+} from "../../infrastructure/serialization.js";
+import { createStorage } from "../../infrastructure/storage/index.js";
 import { loadBackendSetting, saveBackendSetting } from "../../infrastructure/storage/settings.js";
 import { OptionsApp } from "./components/OptionsApp.js";
-import { downloadAsJson, readFileAsJson } from "./fileIO.js";
+
+const downloadAsJson = (data: unknown, filename: string): void => {
+  const blob = new Blob([JSON.stringify(data, null, 2)], { type: "application/json" });
+  const url = URL.createObjectURL(blob);
+  const a = document.createElement("a");
+  a.href = url;
+  a.download = filename;
+  a.click();
+  URL.revokeObjectURL(url);
+};
+
+const readFileAsJson = <T,>(file: File): Promise<T> =>
+  new Promise((resolve, reject) => {
+    const reader = new FileReader();
+    reader.onload = () => {
+      try {
+        resolve(JSON.parse(reader.result as string));
+      } catch {
+        reject(new Error("Invalid JSON"));
+      }
+    };
+    reader.onerror = () => reject(new Error("Failed to read file"));
+    reader.readAsText(file);
+  });
 
 const main = async () => {
   const container = document.querySelector("main");
@@ -16,24 +39,24 @@ const main = async () => {
 
   const [initialBackend, storage] = await Promise.all([loadBackendSetting(), createStorage()]);
 
-  const handleExport = async (): Promise<string> => {
+  const handleExport = async (): Promise<Result<void, string>> => {
     const result = await storage.load();
-    if (!isOk(result)) return "Failed to load vocabulary.";
+    if (!isOk(result)) return err("Failed to load vocabulary.");
     downloadAsJson(serializeVocab(result.value), "pluckvocab-export.json");
-    return "Exported.";
+    return ok(undefined);
   };
 
-  const handleImport = async (file: File): Promise<string> => {
+  const handleImport = async (file: File): Promise<Result<void, string>> => {
     try {
       const loadResult = await storage.load();
-      if (!isOk(loadResult)) return "Failed to load current vocabulary.";
+      if (!isOk(loadResult)) return err("Failed to load current vocabulary.");
       const imported = deserializeVocab(await readFileAsJson<SerializedEntry[]>(file));
       const merged = loadResult.value.merge(imported);
       const saveResult = await storage.save(merged);
-      if (!isOk(saveResult)) return "Failed to save vocabulary.";
-      return "Imported.";
+      if (!isOk(saveResult)) return err("Failed to save vocabulary.");
+      return ok(undefined);
     } catch {
-      return "Invalid file.";
+      return err("Invalid file.");
     }
   };
